update escape calc
